requirements.txt
flask==2.0.3
flask-restful==0.3.9
pillow==8.1.1
tensorflow==2.11.0
pyclamd==0.4.0